add transaction config

+++ src/main/java/com/fcmb/usersecurity/models/User.java
@@ -2,13 +2,12 @@ package com.fcmb.usersecurity.models;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.stereotype.Service;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * A placeholder model to simulate users in the app
@@ -21,5 +20,5 @@ import java.math.BigDecimal;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigDecimal id;
+    private BigInteger id;
 }
